Moving display related helpers to dedicated file.
