+++ resources/assets/components/blocks/CallToActionBlock/CallToActionBlock.js
@@ -6,8 +6,8 @@ import { css } from '@emotion/core';
 import { useQuery } from '@apollo/react-hooks';
 
 import { env } from '../../../helpers/env';
-import { tailwind } from '../../../helpers';
 import ErrorBlock from '../ErrorBlock/ErrorBlock';
+import { tailwind } from '../../../helpers/display';
 import Spinner from '../../artifacts/Spinner/Spinner';
 import PrimaryButton from '../../utilities/Button/PrimaryButton';
