+++ resources/assets/components/utilities/LazyImage.js
@@ -4,7 +4,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import requiredIf from 'react-required-if';
 
-import { EMPTY_IMAGE } from '../../helpers';
+import { EMPTY_IMAGE } from '../../helpers/display';
 
 class LazyImage extends React.Component {
   constructor(props) {
