+++ resources/assets/helpers/index.js
@@ -1,4 +1,4 @@
-/* global window, document, Blob */
+/* global window, Blob */
 
 import { format, getTime, isWithinInterval } from 'date-fns';
 import {
@@ -18,11 +18,6 @@ import {
 import { query } from './url';
 import Debug from '../services/Debug';
 import Sixpack from '../services/Sixpack';
-import tailwindVariables from '../../../tailwind.variables';
-
-// Helper Constants
-export const EMPTY_IMAGE =
-  'data:image/gif;base64,R0lGODlhAQABAIAAAAAAAP///yH5BAEAAAAALAAAAAABAAEAAAIBRAA7';
 
 /**
  * Return a boolean indicating whether the provided argument is an empty string.
@@ -109,35 +104,6 @@ export function withoutTokens(string) {
   const regex = new RegExp(`{[A-Za-z]*}`, 'g');
 
   return string.replace(regex, 'x');
-}
-
-/**
- * Wait until the DOM is ready.
- *
- * @param {Function} fn
- */
-export function ready(fn) {
-  if (document.readyState !== 'loading') {
-    fn();
-  } else {
-    document.addEventListener('DOMContentLoaded', fn);
-  }
-}
-
-/**
- * Prefix a class name or array of class names.
- * @param {String|Array} names
- */
-export function modifiers(...names) {
-  let classes = names;
-
-  if (!Array.isArray(classes)) {
-    classes = [classes];
-  }
-
-  return classes
-    .filter(className => className)
-    .map(className => `-${className}`);
 }
 
 /**
@@ -362,45 +328,6 @@ export function findById(array, compareId) {
 }
 
 /**
- * Open a dialog and run a callback when it closes.
- *
- * @param {String} href
- * @param {Function} callback
- * @param {Number} height
- * @param {Number} width
- */
-export function openDialog(href, callback, width = 550, height = 420) {
-  const winHeight = window.screen.height;
-  const winWidth = window.screen.width;
-
-  const left = Math.round(winWidth / 2 - width / 2);
-  let top = 0;
-
-  if (winHeight > height) {
-    top = Math.round(winHeight / 2 - height / 2);
-  }
-
-  const dialog = window.open(
-    href,
-    'intent',
-    `scrollbars=yes,resizable=yes,toolbar=no,location=yes,width=${width},height=${height},left=${left},top=${top}`,
-  );
-
-  let interval;
-
-  const check = () => {
-    if (dialog.closed) {
-      clearInterval(interval);
-      callback();
-    }
-  };
-
-  if (callback) {
-    interval = setInterval(check, 1000);
-  }
-}
-
-/**
  * Report a caught error to New Relic.
  *
  * @param {Error|string}  error
@@ -498,49 +425,6 @@ export function getScholarshipAffiliateLabel() {
   // If the utm_source contains 'scholarship', we assume this visit to be a referral from a
   // scholarship affiliate and return the affiliate's UTM label.
   return isScholarshipAffiliateReferral() ? utmLabel : null;
-}
-
-/**
- * Get specified theme setting from the resolved Tailwind configuration object.
- *
- * @param {String} themeSetting
- */
-export function tailwind(themeSetting) {
-  if (!isString(themeSetting)) {
-    throw new Error(
-      'Please specify a theme setting as a string to retrieve from Tailwind.',
-    );
-  }
-
-  const setting = get(tailwindVariables, themeSetting, null);
-
-  if (!setting) {
-    console.error(
-      `The ${themeSetting} setting specified was not found in the Tailwind theme configuration.`,
-    );
-  }
-
-  return setting;
-}
-
-/**
- * Toggle the specified class on the given target element
- * when the button element is clicked or touched.
- *
- * @param  {Element} button
- * @param  {Element} target
- * @param  {String} toggleClass
- */
-export function toggleClassHandler(button, target, toggleClass) {
-  if (!button || !target) {
-    return;
-  }
-
-  function clickHandler() {
-    target.classList.toggle(toggleClass);
-  }
-
-  button.addEventListener('mousedown', clickHandler, false);
 }
 
 /**
