+++ resources/assets/helpers/twitter.js
@@ -1,5 +1,5 @@
-import { openDialog } from '.';
 import { makeUrl } from './url';
+import { openDialog } from './display';
 import { EVENT_CATEGORIES, trackAnalyticsEvent } from './analytics';
 
 /**
